fiexd projections error

+++ src/main/java/com/examplerm/rmdemo/entities/projections/ArtistProjection.java
@@ -6,6 +6,6 @@ public interface ArtistProjection {
 
     String getName();
 
-    Long getListener();
+    Integer getListener();
 
 }
